refactor: 쿠폰 관리 컴포넌트 생성 및 적용
