+++ src/origin/pages/AdminPage.tsx
@@ -1,9 +1,8 @@
 import { useState } from 'react';
-import { Discount, Product } from '../../types';
+import { Discount, Product } from '../../types.ts';
 import { useProductStore } from '../stores/useProductStore.ts';
-import { useCouponStore } from '../stores/useCouponStore.ts';
 import { useNewProduct } from '../hooks/useNewProduct.ts';
-import { useNewCoupon } from '../hooks/useNewCoupon.ts';
+import CouponManagement from '../components/CouponManagement.tsx';
 
 export const AdminPage = () => {
   const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
@@ -15,10 +14,8 @@ export const AdminPage = () => {
   const [showNewProductForm, setShowNewProductForm] = useState(false);
 
   const { newProduct, initNewProduct, updateNewProduct } = useNewProduct();
-  const { newCoupon, initNewCoupon, updateNewCoupon } = useNewCoupon();
 
   const { products, updateProduct, addProduct } = useProductStore();
-  const { coupons, addCoupon } = useCouponStore();
 
   const toggleProductAccordion = (productId: string) => {
     setOpenProductIds((prev) => {
@@ -93,11 +90,6 @@ export const AdminPage = () => {
       updateProduct(newProduct);
       setEditingProduct(newProduct);
     }
-  };
-
-  const handleAddCoupon = () => {
-    addCoupon(newCoupon);
-    initNewCoupon();
   };
 
   const handleAddNewProduct = () => {
@@ -336,82 +328,7 @@ export const AdminPage = () => {
             ))}
           </div>
         </div>
-        <div>
-          <h2 className="text-2xl font-semibold mb-4">쿠폰 관리</h2>
-          <div className="bg-white p-4 rounded shadow">
-            <div className="space-y-2 mb-4">
-              <input
-                type="text"
-                placeholder="쿠폰 이름"
-                value={newCoupon.name}
-                onChange={(e) =>
-                  updateNewCoupon({ ...newCoupon, name: e.target.value })
-                }
-                className="w-full p-2 border rounded"
-              />
-              <input
-                type="text"
-                placeholder="쿠폰 코드"
-                value={newCoupon.code}
-                onChange={(e) =>
-                  updateNewCoupon({ ...newCoupon, code: e.target.value })
-                }
-                className="w-full p-2 border rounded"
-              />
-              <div className="flex gap-2">
-                <select
-                  value={newCoupon.discountType}
-                  onChange={(e) =>
-                    updateNewCoupon({
-                      ...newCoupon,
-                      discountType: e.target.value as 'amount' | 'percentage'
-                    })
-                  }
-                  className="w-full p-2 border rounded"
-                >
-                  <option value="amount">금액(원)</option>
-                  <option value="percentage">할인율(%)</option>
-                </select>
-                <input
-                  type="number"
-                  placeholder="할인 값"
-                  value={newCoupon.discountValue}
-                  onChange={(e) =>
-                    updateNewCoupon({
-                      ...newCoupon,
-                      discountValue: parseInt(e.target.value)
-                    })
-                  }
-                  className="w-full p-2 border rounded"
-                />
-              </div>
-              <button
-                onClick={handleAddCoupon}
-                className="w-full bg-green-500 text-white p-2 rounded hover:bg-green-600"
-              >
-                쿠폰 추가
-              </button>
-            </div>
-            <div>
-              <h3 className="text-lg font-semibold mb-2">현재 쿠폰 목록</h3>
-              <div className="space-y-2">
-                {coupons.map((coupon, index) => (
-                  <div
-                    key={index}
-                    data-testid={`coupon-${index + 1}`}
-                    className="bg-gray-100 p-2 rounded"
-                  >
-                    {coupon.name} ({coupon.code}):
-                    {coupon.discountType === 'amount'
-                      ? `${coupon.discountValue}원`
-                      : `${coupon.discountValue}%`}{' '}
-                    할인
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </div>
+        <CouponManagement />
       </div>
     </div>
   );
